Display a welcome video to new accounts
FIXME: we'll need a beter way, especially because this won't show up for
new users on busy accounts. To avoid seeing it blip everytime you
paginate it's set to `display: none` if there are enough events to
paginate. So people can scroll down to watch again if they only have a
few events.